requirements.txt
flask
h5py
h5json
pillow
numpy
